+++ cameo-server/src/main/java/fr/ill/ics/cameo/manager/PortManager.java
@@ -17,6 +17,7 @@
 package fr.ill.ics.cameo.manager;
 
 import java.util.HashMap;
+import java.util.HashSet;
 
 public final class PortManager {
 
@@ -32,13 +33,24 @@ public final class PortManager {
 	 * State of an reserved port.
 	 */
 	private static class State {
+		
 		Status status = Status.ASSIGNED;
+		Integer applicationId = null;
+		
+		public State(int applicationId) {
+			this.applicationId = applicationId;
+		}
 	}
 	
 	/**
 	 * Map of the reserved ports.
 	 */
 	private HashMap<Integer, State> reservedPorts = new HashMap<Integer, State>();
+	
+	/**
+	 * Map of the application ports.
+	 */
+	private HashMap<Integer, HashSet<Integer>> applicationPorts = new HashMap<Integer, HashSet<Integer>>();
 	
 	/**
 	 * Constructor.
@@ -64,21 +76,37 @@ public final class PortManager {
 	}
 
 	/**
+	 * Get the application ports. Create the entry if it does not exist.
+	 * @param applicationId the application id
+	 * @return the application ports
+	 */
+	private HashSet<Integer> getApplicationPorts(int applicationId) {
+		
+		if (!applicationPorts.containsKey(applicationId)) {
+			applicationPorts.put(applicationId, new HashSet<Integer>());
+		}
+		return applicationPorts.get(applicationId);
+	}
+	
+	/**
 	 * Request a non-reserved port. However the port can be unavailable because another application opened it.
 	 * @return a port
 	 */
-	public int requestPort() {
+	public int requestPort(int applicationId) {
 
 		// Loop from the base port.
 		int port = basePort;
 		while (true) {
 			
+			// Continue to iterate if the port is already reserved.
 			if (reservedPorts.containsKey(port)) {
 				port++;
 			}
 			else {
-				// Found a port.
-				reservedPorts.put(port, new State());
+				// Found a port. Add it in the reserved ports and application ports.
+				reservedPorts.put(port, new State(applicationId));
+				getApplicationPorts(applicationId).add(port);
+				
 				break;
 			}
 		}
@@ -91,7 +119,33 @@ public final class PortManager {
 	 * @param port the port to remove
 	 */
 	public void removePort(int port) {
-		reservedPorts.remove(port);
+		
+		// Remove the port from the reserved list.
+		State state = reservedPorts.remove(port);
+		
+		// Remove the state from application.
+		if (state != null 
+				&& state.status == Status.ASSIGNED
+				&& state.applicationId != null) {
+			getApplicationPorts(state.applicationId).remove(port);
+		}
+	}
+	
+	/**
+	 * Remove the ports of the application. 
+	 * @param applicationId the application id
+	 */
+	public void removeApplication(int applicationId) {
+		
+		// Remove the ports of the application.
+		HashSet<Integer> ports = getApplicationPorts(applicationId);
+		
+		for (Integer port : ports) {
+			reservedPorts.remove(port);
+		}
+		
+		// Remove the application.
+		applicationPorts.remove(applicationId);
 	}
 	
 	/**
@@ -99,8 +153,20 @@ public final class PortManager {
 	 * @param port the port to set unavailable
 	 */
 	public void setPortUnavailable(int port) {
+		
 		if (reservedPorts.containsKey(port)) {
-			reservedPorts.get(port).status = Status.UNAVAILABLE;
+			State state = reservedPorts.get(port);
+			
+			// Set it unavailable.
+			state.status = Status.UNAVAILABLE;
+						
+			// Remove from the application.
+			if (state.applicationId != null) {
+				getApplicationPorts(state.applicationId).remove(port);
+			}
+			
+			// Set the application id to null;
+			state.applicationId = null;
 		}
 	}