+++ cameo-server/CHANGELOG.md
@@ -14,6 +14,7 @@
 * Enabled multiple=yes and stream=yes.
 * Added connect with id request.
 * Added sync stream request. Moved the creation of the stream thread at an earlier location.
+* The class PortManager is responsible for managing the ports.
 
 0.1.8
 -----